requirements.txt
bitarray
pycrypto
